refactor: side panel component
